Update Merkly contract and OFT contract

+++ contracts/OFT.sol
@@ -25,18 +25,27 @@ contract OFT is OFTCore, ERC20, IOFT {
         return address(this);
     }
 
-    function circulatingSupply() public view virtual override returns (uint) {
+    function circulatingSupply() public view virtual override returns (uint256) {
         return totalSupply();
     }
 
-    function _debitFrom(address _from, uint16, bytes memory, uint _amount) internal virtual override returns (uint) {
+    function _debitFrom(
+        address _from,
+        uint16,
+        bytes memory,
+        uint256 _amount
+    ) internal virtual override returns (uint256) {
         address spender = _msgSender();
         if (_from != spender) _spendAllowance(_from, spender, _amount);
         _burn(_from, _amount);
         return _amount;
     }
 
-    function _creditTo(uint16, address _toAddress, uint _amount) internal virtual override returns (uint) {
+    function _creditTo(
+        uint16,
+        address _toAddress,
+        uint256 _amount
+    ) internal virtual override returns (uint256) {
         _mint(_toAddress, _amount);
         return _amount;
     }
