+++ contracts/OFTCore.sol
@@ -8,7 +8,7 @@ import "./interfaces/IOFTCore.sol";
 abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
     using BytesLib for bytes;
 
-    uint public constant NO_EXTRA_GAS = 0;
+    uint256 public constant NO_EXTRA_GAS = 0;
 
     // packet type
     uint16 public constant PT_SEND = 0;
@@ -24,10 +24,10 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
     function estimateSendFee(
         uint16 _dstChainId,
         bytes calldata _toAddress,
-        uint _amount,
+        uint256 _amount,
         bool _useZro,
         bytes calldata _adapterParams
-    ) public view virtual override returns (uint nativeFee, uint zroFee) {
+    ) public view virtual override returns (uint256 nativeFee, uint256 zroFee) {
         // mock the payload for sendFrom()
         bytes memory payload = abi.encode(PT_SEND, _toAddress, _amount);
         return lzEndpoint.estimateFees(_dstChainId, address(this), payload, _useZro, _adapterParams);
@@ -38,7 +38,7 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
         bytes memory payload,
         bool _useZro,
         bytes memory _adapterParams
-    ) public view virtual returns (uint nativeFee, uint zroFee) {
+    ) public view virtual returns (uint256 nativeFee, uint256 zroFee) {
         return lzEndpoint.estimateFees(_dstChainId, address(this), payload, _useZro, _adapterParams);
     }
 
@@ -46,7 +46,7 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
         address _from,
         uint16 _dstChainId,
         bytes calldata _toAddress,
-        uint _amount,
+        uint256 _amount,
         address payable _refundAddress,
         address _zroPaymentAddress,
         bytes calldata _adapterParams
@@ -81,18 +81,18 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
         address _from,
         uint16 _dstChainId,
         bytes memory _toAddress,
-        uint _amount,
+        uint256 _amount,
         address payable _refundAddress,
         address _zroPaymentAddress,
         bytes memory _adapterParams
     ) internal virtual {
         _checkAdapterParams(_dstChainId, PT_SEND, _adapterParams, NO_EXTRA_GAS);
 
-        uint amount = _debitFrom(_from, _dstChainId, _toAddress, _amount);
+        uint256 amount = _debitFrom(_from, _dstChainId, _toAddress, _amount);
 
         bytes memory lzPayload = abi.encode(PT_SEND, _toAddress, amount);
 
-        (uint nativeFee, ) = estimateSendFee2(_dstChainId, lzPayload, false, _adapterParams);
+        (uint256 nativeFee, ) = estimateSendFee2(_dstChainId, lzPayload, false, _adapterParams);
         require(msg.value >= nativeFee, "Not enough gas to send");
 
         _lzSend(_dstChainId, lzPayload, _refundAddress, _zroPaymentAddress, _adapterParams, nativeFee);
@@ -100,8 +100,13 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
         emit SendToChain(_dstChainId, _from, _toAddress, amount);
     }
 
-    function _sendAck(uint16 _srcChainId, bytes memory, uint64, bytes memory _payload) internal virtual {
-        (, bytes memory toAddressBytes, uint amount) = abi.decode(_payload, (uint16, bytes, uint));
+    function _sendAck(
+        uint16 _srcChainId,
+        bytes memory,
+        uint64,
+        bytes memory _payload
+    ) internal virtual {
+        (, bytes memory toAddressBytes, uint256 amount) = abi.decode(_payload, (uint16, bytes, uint256));
 
         address to = toAddressBytes.toAddress(0);
 
@@ -113,7 +118,7 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
         uint16 _dstChainId,
         uint16 _pkType,
         bytes memory _adapterParams,
-        uint _extraGas
+        uint256 _extraGas
     ) internal virtual {
         if (useCustomAdapterParams) {
             _checkGasLimit(_dstChainId, _pkType, _adapterParams, _extraGas);
@@ -126,8 +131,12 @@ abstract contract OFTCore is NonblockingLzApp, ERC165, IOFTCore {
         address _from,
         uint16 _dstChainId,
         bytes memory _toAddress,
-        uint _amount
-    ) internal virtual returns (uint);
+        uint256 _amount
+    ) internal virtual returns (uint256);
 
-    function _creditTo(uint16 _srcChainId, address _toAddress, uint _amount) internal virtual returns (uint);
+    function _creditTo(
+        uint16 _srcChainId,
+        address _toAddress,
+        uint256 _amount
+    ) internal virtual returns (uint256);
 }
